Salary change and routes change

+++ models/salary.js
@@ -1,10 +1,10 @@
 module.exports = function(sequelize, DataTypes) {
 
     var Salary = sequelize.define("Salary", {
-        pay_rate: DataTypes.INTEGER,
+        pay_rate: DataTypes.DECIMAL(5,2),
         end_date: DataTypes.DATEONLY,
      });
-  
+
      Salary.associate = function(models) {
         Salary.belongsTo(models.Employee, {
             foreignKey: {
@@ -12,7 +12,6 @@ module.exports = function(sequelize, DataTypes) {
             }
           });
         };
-    
+
     return Salary;
   };
-
